Move chunker from inference to core
