+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/inference/chunking/SentenceBoundaryChunker.java
@@ -5,13 +5,12 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.inference.chunking;
+package org.elasticsearch.xpack.core.inference.chunking;
 
 import com.ibm.icu.text.BreakIterator;
 
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.inference.ChunkingSettings;
-import org.elasticsearch.xpack.core.inference.chunking.SentenceBoundaryChunkingSettings;
 
 import java.util.ArrayList;
 import java.util.List;
