+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/inference/chunking/ChunkerBuilderTests.java
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.inference.chunking;
+package org.elasticsearch.xpack.core.inference.chunking;
 
 import org.elasticsearch.inference.ChunkingStrategy;
 import org.elasticsearch.test.ESTestCase;
