+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/inference/chunking/RecursiveChunkerTests.java
@@ -5,13 +5,10 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.inference.chunking;
+package org.elasticsearch.xpack.core.inference.chunking;
 
 import org.elasticsearch.inference.ChunkingSettings;
 import org.elasticsearch.test.ESTestCase;
-import org.elasticsearch.xpack.core.inference.chunking.RecursiveChunkingSettings;
-import org.elasticsearch.xpack.core.inference.chunking.SentenceBoundaryChunkingSettings;
-import org.elasticsearch.xpack.core.inference.chunking.SeparatorGroup;
 
 import java.util.ArrayList;
 import java.util.List;
