+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/inference/chunking/WordBoundaryChunkerTests.java
@@ -5,15 +5,12 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.inference.chunking;
+package org.elasticsearch.xpack.core.inference.chunking;
 
 import com.ibm.icu.text.BreakIterator;
 
 import org.elasticsearch.inference.ChunkingSettings;
 import org.elasticsearch.test.ESTestCase;
-import org.elasticsearch.xpack.core.inference.chunking.NoneChunkingSettings;
-import org.elasticsearch.xpack.core.inference.chunking.SentenceBoundaryChunkingSettings;
-import org.elasticsearch.xpack.core.inference.chunking.WordBoundaryChunkingSettings;
 import org.hamcrest.Matchers;
 
 import java.util.List;
