+++ x-pack/plugin/inference/build.gradle
@@ -41,8 +41,6 @@ dependencies {
   testImplementation project(':x-pack:plugin:ml')
   clusterPlugins project(':x-pack:plugin:inference:qa:test-service-plugin')
 
-  api "com.ibm.icu:icu4j:${versions.icu4j}"
-
   runtimeOnly 'com.google.guava:guava:32.0.1-jre'
   implementation 'com.google.code.gson:gson:2.10'
   implementation "com.google.protobuf:protobuf-java-util:${versions.protobuf}"
