+++ x-pack/plugin/inference/qa/test-service-plugin/src/main/java/org/elasticsearch/xpack/inference/mock/AbstractTestInferenceService.java
@@ -26,8 +26,8 @@ import org.elasticsearch.inference.TaskSettings;
 import org.elasticsearch.inference.TaskType;
 import org.elasticsearch.xcontent.XContentBuilder;
 import org.elasticsearch.xpack.core.inference.chunking.WordBoundaryChunkingSettings;
-import org.elasticsearch.xpack.inference.chunking.NoopChunker;
-import org.elasticsearch.xpack.inference.chunking.WordBoundaryChunker;
+import org.elasticsearch.xpack.core.inference.chunking.NoopChunker;
+import org.elasticsearch.xpack.core.inference.chunking.WordBoundaryChunker;
 
 import java.io.IOException;
 import java.util.ArrayList;
