+++ x-pack/plugin/inference/src/main/java/module-info.java
@@ -18,7 +18,6 @@ module org.elasticsearch.inference {
     requires org.apache.httpcomponents.httpcore.nio;
     requires org.apache.lucene.core;
     requires org.apache.lucene.join;
-    requires com.ibm.icu;
     requires com.google.auth.oauth2;
     requires com.google.auth;
     requires com.google.api.client;
@@ -44,7 +43,6 @@ module org.elasticsearch.inference {
     exports org.elasticsearch.xpack.inference.services;
     exports org.elasticsearch.xpack.inference;
     exports org.elasticsearch.xpack.inference.action.task;
-    exports org.elasticsearch.xpack.inference.chunking;
     exports org.elasticsearch.xpack.inference.telemetry;
 
     provides org.elasticsearch.features.FeatureSpecification with org.elasticsearch.xpack.inference.InferenceFeatures;
